+++ clients/java/src/main/java/io/lakefs/clients/sdk/model/CommitRecordCreation.java
@@ -62,7 +62,7 @@ public class CommitRecordCreation {
 
   public static final String SERIALIZED_NAME_VERSION = "version";
   @SerializedName(SERIALIZED_NAME_VERSION)
-  private Long version;
+  private Integer version;
 
   public static final String SERIALIZED_NAME_COMMITTER = "committer";
   @SerializedName(SERIALIZED_NAME_COMMITTER)
@@ -120,7 +120,7 @@ public class CommitRecordCreation {
   }
 
 
-  public CommitRecordCreation version(Long version) {
+  public CommitRecordCreation version(Integer version) {
     
     this.version = version;
     return this;
@@ -128,15 +128,17 @@ public class CommitRecordCreation {
 
    /**
    * version of the commit record
+   * minimum: 0
+   * maximum: 1
    * @return version
   **/
   @javax.annotation.Nonnull
-  public Long getVersion() {
+  public Integer getVersion() {
     return version;
   }
 
 
-  public void setVersion(Long version) {
+  public void setVersion(Integer version) {
     this.version = version;
   }
